pyproject.toml
[tool.poetry]
name = "shiki-recs"
version = "0.1.0"
description = ""
authors = ["Annovid <nil.nil532@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.12"
alembic = "^1.13.1"
sqlalchemy = "^2.0.29"
ruff = "^0.3.6"
shikimori-api = "^1.0.1"
pydantic = "^2.9.0"
pytest = "^8.3.3"
python-telegram-bot = "^21.6"
telebot = "^0.0.5"
bs4 = "^0.0.2"
pydantic-settings = "^2.6.0"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79